提交近期所有变更：修复多个bug和功能改进

+++ api/src/main/java/io/github/guojiaxing1995/easyJmeter/service/impl/ProjectServiceImpl.java
@@ -74,6 +74,6 @@ public class ProjectServiceImpl implements ProjectService {
     public Integer getCaseCount(Integer id) {
         QueryWrapper<CaseDO> queryWrapper = new QueryWrapper<>();
         queryWrapper.eq("project", id);
-        return caseMapper.selectCount(queryWrapper);
+        return caseMapper.selectCount(queryWrapper).intValue();
     }
 }
